requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
certifi==2025.11.12
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.121.3
fonttools==4.60.1
h11==0.16.0
idna==3.11
imbalanced-learn==0.14.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
pyasn1==0.6.1
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
xgboost==3.1.2
